rec-mode: fix error when saving a newly created file
Hi,

Perhaps this is the same issue as TODO 14.  I get a "Wrong type
argument: number-or-marker-p, nil" error, when I open a non-existing
file and finish editing.  In that case, an erroneous value is set to
rec-buffer-descriptors:

((descriptor error: #<marker in no buffer>)
 (descriptor cannot #<marker in no buffer>)
 (descriptor read #<marker in no buffer>)
 (descriptor file #<marker in no buffer>)
 (descriptor /home/ueno/recutils/foo\.rec #<marker in no buffer>))

So I guess this could be fixed by feeding data to recinf from the buffer
instead of the file.  Here is a patch.

2012-04-09  Daiki Ueno  <ueno@unixuser.org>

	rec-mode: fix error when rec-finish-editing on newly created file.
	* etc/rec-mode.el (rec-update-buffer-descriptors): Don't assume
	that the current buffer is already saved on the disk.

>From 5db6934c2ab84080b72bfc7663db9495640dbfda Mon Sep 17 00:00:00 2001
From: Daiki Ueno <ueno@unixuser.org>
Date: Mon, 9 Apr 2012 11:29:10 +0900
Subject: [PATCH] rec-mode: fix error when rec-finish-editing on newly created
 file.