Documentation: Primary and foreign keys
There were several references to primary and foreign keys scattered
around the manual.  To some extent this is unavoidable.  However, this
change puts the main discussion of the subject into one place, and
changes the other instances to mention the subject in a cursorary
fashion and a cross reference to the main discussion.